+++ packages/lumx-vue/src/index.ts
@@ -18,6 +18,7 @@ export * from './components/icon';
 export * from './components/input-helper';
 export * from './components/input-label';
 export * from './components/link';
+export * from './components/link-preview';
 export * from './components/list';
 export * from './components/message';
 export * from './components/mosaic';